datagen/pyproject.toml
[project]
name = "zebra-evol"
version = "0.1.0"
description = "Evolution of zebra/Einstein logic puzzles using CAMEL's EvolInstruct pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.8.0",
    "pyarrow>=10.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-constraint>=1.4.0",
    "camel-ai[all]>=0.2.74",
    "faiss-cpu==1.10.0",    # compatible macOS arm64 wheel
    "traceroot==0.0.4a5",    # force pre-release so uv won't block it
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.1",
    "requests>=2.28.0"
]
